Move command framework into kyori module
