--- minecraft/commands/build.gradle.kts
@@ -1,17 +1,0 @@
-plugins {
-    `config-publish`
-    `config-java`
-}
-
-project.group = "${rootProject.name}.commands"
-
-repositories {
-    maven("https://libraries.minecraft.net")
-}
-
-dependencies {
-    compileOnly(libs.bundles.adventure)
-    compileOnly(libs.brigadier)
-
-    api(project(":fusion-kyori"))
-}
