--- minecraft/commands/src/main/java/com/ryderbelserion/fusion/commands/api/annotations/Flower.java
@@ -1,14 +1,0 @@
-package com.ryderbelserion.fusion.commands.api.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.METHOD})
-public @interface Flower {
-
-    String desc() default "";
-
-}
