--- minecraft/commands/src/main/java/com/ryderbelserion/fusion/commands/api/annotations/other/Permission.java
@@ -1,14 +1,0 @@
-package com.ryderbelserion.fusion.commands.api.annotations.other;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.TYPE, ElementType.METHOD})
-public @interface Permission {
-
-    String permission() default "";
-
-}
