--- minecraft/commands/src/main/java/com/ryderbelserion/fusion/commands/api/annotations/other/Requirement.java
@@ -1,10 +1,0 @@
-package com.ryderbelserion.fusion.commands.api.annotations.other;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.ANNOTATION_TYPE)
-public @interface Requirement {}
