--- minecraft/commands/src/main/java/com/ryderbelserion/fusion/commands/api/annotations/subs/Branch.java
@@ -1,16 +1,0 @@
-package com.ryderbelserion.fusion.commands.api.annotations.subs;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.TYPE, ElementType.METHOD, ElementType.ANNOTATION_TYPE})
-public @interface Branch { // tree i.e. /fusion help
-
-    String value() default "";
-
-    String desc() default "";
-
-}
