--- minecraft/commands/src/main/java/com/ryderbelserion/fusion/commands/api/annotations/subs/Leaf.java
@@ -1,16 +1,0 @@
-package com.ryderbelserion.fusion.commands.api.annotations.subs;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.TYPE, ElementType.METHOD})
-public @interface Leaf { // tree i.e. /fusion help
-
-    String value() default "";
-
-    String desc() default "";
-
-}
