+++ minecraft/kyori/build.gradle.kts
@@ -5,8 +5,13 @@ plugins {
 
 project.group = "${rootProject.name}.kyori"
 
+repositories {
+    maven("https://libraries.minecraft.net")
+}
+
 dependencies {
     compileOnly(libs.bundles.adventure)
+    compileOnly(libs.brigadier)
 
     api(project(":fusion-core"))
 }
