+++ minecraft/kyori/src/main/java/com/ryderbelserion/fusion/kyori/commands/api/annotations/Flower.java
@@ -1,0 +1,14 @@
+package com.ryderbelserion.fusion.kyori.commands.api.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.METHOD})
+public @interface Flower {
+
+    String desc() default "";
+
+}
