+++ minecraft/kyori/src/main/java/com/ryderbelserion/fusion/kyori/commands/api/annotations/other/Requirement.java
@@ -1,0 +1,10 @@
+package com.ryderbelserion.fusion.kyori.commands.api.annotations.other;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.ANNOTATION_TYPE)
+public @interface Requirement {}
