+++ minecraft/kyori/src/main/java/com/ryderbelserion/fusion/kyori/commands/api/senders/objects/SenderExtension.java
@@ -1,0 +1,20 @@
+package com.ryderbelserion.fusion.kyori.commands.api.senders.objects;
+
+import com.ryderbelserion.fusion.kyori.commands.api.senders.results.ValidationResult;
+import org.jetbrains.annotations.NotNull;
+import java.util.Set;
+
+public interface SenderExtension<S> {
+
+    @NotNull ValidationResult<?> validate(final @NotNull Class<?> target, final @NotNull S sender);
+
+    @NotNull Set<Class<? extends S>> getSenders();
+
+    default ValidationResult<?> invalid(String message) {
+        return new ValidationResult.Invalid<>(message);
+    }
+
+    default ValidationResult<?> valid() {
+        return new ValidationResult.Valid<>();
+    }
+}
