+++ minecraft/kyori/src/main/java/com/ryderbelserion/fusion/kyori/commands/api/senders/results/ValidationResult.java
@@ -1,0 +1,22 @@
+package com.ryderbelserion.fusion.kyori.commands.api.senders.results;
+
+import org.jetbrains.annotations.NotNull;
+
+public interface ValidationResult<E> {
+
+    class Valid<E> implements ValidationResult<E> {}
+
+    class Ignore<E> implements ValidationResult<E> {}
+
+    record Invalid<E> (E message) implements ValidationResult<E> {
+
+        public Invalid(final @NotNull E message) {
+            this.message = message;
+        }
+
+        @Override
+        public @NotNull E message() {
+            return this.message;
+        }
+    }
+}
