+++ minecraft/paper/example/build.gradle.kts
@@ -10,8 +10,7 @@ repositories {
 }
 
 dependencies {
-    implementation(project(":fusion-commands"))
-    //implementation(project(":fusion-paper"))
+    implementation(project(":fusion-paper"))
 }
 
 tasks {
