--- minecraft/paper/example/src/main/java/com/ryderbelserion/fusion/api/PaperCommandManager.java
@@ -1,44 +1,0 @@
-package com.ryderbelserion.fusion.api;
-
-import com.mojang.brigadier.builder.LiteralArgumentBuilder;
-import com.ryderbelserion.fusion.api.senders.CommandSender;
-import com.ryderbelserion.fusion.api.senders.PlayerSender;
-import com.ryderbelserion.fusion.commands.CommandManager;
-import com.ryderbelserion.fusion.commands.api.objects.AbstractCommand;
-import com.ryderbelserion.fusion.commands.api.TreeProcessor;
-import io.papermc.paper.command.brigadier.CommandSourceStack;
-import io.papermc.paper.plugin.lifecycle.event.LifecycleEventManager;
-import io.papermc.paper.plugin.lifecycle.event.types.LifecycleEvents;
-import org.bukkit.plugin.Plugin;
-import org.bukkit.plugin.java.JavaPlugin;
-import org.jetbrains.annotations.NotNull;
-
-public class PaperCommandManager extends CommandManager<CommandSourceStack> {
-
-    private final LifecycleEventManager<Plugin> eventManager;
-
-    public PaperCommandManager(@NotNull final JavaPlugin plugin) {
-        this.eventManager = plugin.getLifecycleManager();
-
-        addSender(new CommandSender());
-        addSender(new PlayerSender());
-    }
-
-    @Override
-    public boolean hasPermission(@NotNull CommandSourceStack context, @NotNull final String permission) {
-        return context.getSender().hasPermission(permission);
-    }
-
-    @Override
-    public void init(@NotNull final String key) {
-        if (!this.commands.containsKey(key)) return;
-
-        final AbstractCommand origin = this.commands.get(key);
-
-        final TreeProcessor processor = origin.getProcessor();
-
-        final LiteralArgumentBuilder<CommandSourceStack> builder = processor.getBuilder();
-
-        this.eventManager.registerEventHandler(LifecycleEvents.COMMANDS, event -> event.registrar().register(builder.build(), processor.getDescription()));
-    }
-}
