+++ minecraft/paper/example/src/main/java/com/ryderbelserion/fusion/commands/types/SubCommand.java
@@ -1,9 +1,9 @@
 package com.ryderbelserion.fusion.commands.types;
 
-import com.ryderbelserion.fusion.commands.api.annotations.Flower;
-import com.ryderbelserion.fusion.commands.api.annotations.other.Permission;
-import com.ryderbelserion.fusion.commands.api.annotations.subs.Branch;
-import com.ryderbelserion.fusion.commands.api.annotations.subs.Leaf;
+import com.ryderbelserion.fusion.kyori.commands.api.annotations.Flower;
+import com.ryderbelserion.fusion.kyori.commands.api.annotations.other.Permission;
+import com.ryderbelserion.fusion.kyori.commands.api.annotations.subs.Branch;
+import com.ryderbelserion.fusion.kyori.commands.api.annotations.subs.Leaf;
 
 @Branch(value = "sub")
 @Permission(permission = "fusion.sub")
