+++ minecraft/paper/src/main/java/com/ryderbelserion/fusion/paper/commands/PaperCommandManager.java
@@ -1,0 +1,51 @@
+package com.ryderbelserion.fusion.paper.commands;
+
+import com.mojang.brigadier.builder.LiteralArgumentBuilder;
+import com.ryderbelserion.fusion.paper.commands.extensions.PaperSenderExtension;
+import com.ryderbelserion.fusion.kyori.commands.CommandManager;
+import com.ryderbelserion.fusion.kyori.commands.api.objects.AbstractCommand;
+import com.ryderbelserion.fusion.kyori.commands.api.TreeProcessor;
+import io.papermc.paper.command.brigadier.CommandSourceStack;
+import io.papermc.paper.plugin.lifecycle.event.LifecycleEventManager;
+import io.papermc.paper.plugin.lifecycle.event.types.LifecycleEvents;
+import org.bukkit.plugin.Plugin;
+import org.bukkit.plugin.java.JavaPlugin;
+import org.jetbrains.annotations.NotNull;
+
+public class PaperCommandManager extends CommandManager<CommandSourceStack> {
+
+    private final LifecycleEventManager<Plugin> eventManager;
+
+    public PaperCommandManager(@NotNull final JavaPlugin plugin) {
+        this.eventManager = plugin.getLifecycleManager();
+    }
+
+    private PaperSenderExtension senderExtension;
+
+    @Override
+    public boolean hasPermission(@NotNull CommandSourceStack context, @NotNull final String permission) {
+        return context.getSender().hasPermission(permission);
+    }
+
+    @Override
+    public void post(@NotNull final String key) {
+        if (!this.commands.containsKey(key)) return;
+
+        final AbstractCommand origin = this.commands.get(key);
+
+        final TreeProcessor processor = origin.getProcessor();
+
+        final LiteralArgumentBuilder<CommandSourceStack> builder = processor.getBuilder();
+
+        this.eventManager.registerEventHandler(LifecycleEvents.COMMANDS, event -> event.registrar().register(builder.build(), processor.getDescription()));
+    }
+
+    @Override
+    public void init() {
+        this.senderExtension = new PaperSenderExtension(this);
+    }
+
+    public @NotNull final PaperSenderExtension getSenderExtension() {
+        return this.senderExtension;
+    }
+}
