+++ settings.gradle.kts
@@ -10,7 +10,6 @@ rootProject.name = "Fusion"
 listOf(
     "minecraft/paper/example" to "example",
 
-    "minecraft/commands" to "commands",
     "minecraft/paper" to "paper",
     "minecraft/kyori" to "kyori",
